-Исправлен рендеринг ui элементов для отладки. -Исправлены лишние вызовы рендеринга шрифтов. -Исправлен рендеринг партиклов на позициях player_hud. -Исправлен hud fov убрана настройка. -Исправлена коррекция fire_point fire_dir и тд для худа. -Исправлена отладка player_hud. -Исправлены некоторые критические ошибки в рендеринге. -Переделана система инерции player_hud. -Немного рефакторинга player_hud. -Обновлена беблиотека freetype для рендеринга шрифтов. -Исправлены критические ошибки в отладке связанные с пустыми xr_vector. -Адаптирован xr_resource под многопоточность.

+++ src/xrEngine/device.h
@@ -248,7 +248,7 @@ public:
 	// Mode control
 	void DumpFlags();
 	IC CTimer_paused* GetTimerGlobal() { return &TimerGlobal; }
-	u32 TimerAsync() { return TimerGlobal.GetElapsed_ms(); }
+	size_t TimerAsync() { return (size_t)TimerGlobal.GetElapsed_ms(); }
 	// Creation & Destroying
 	void Create();
 	void Run();
